Fix Windows test, where the dtype was inferred as int32.

+++ python/test_xnd.py
@@ -2770,7 +2770,7 @@ class TestBuffer(XndTestCase):
         x = np.array([[[0,1,2],
                        [3,4,5]],
                      [[6,7,8],
-                      [9,10,11]]])
+                      [9,10,11]]], dtype="int64")
 
         y = xnd.unsafe_from_data(obj=x, type="12 * int64")
         np.testing.assert_equal(y, x.reshape(12))
